Add semaphore build pipeline
This introduces semaphore (2.0) ci/cd pipeline that we will use for
fiaas-deploy-daemon building and release process going forward.
Previously we have used an internal travis instance and private
pipelines for building and releasing fiaas-deploy-daemon. With
Semaphore we can simplify the workflow and make the release process
use public tooling.

The current pipeline will build the project and run unit-tests, build,
tag and publish a docker image, and publish artifacts and meta-data that
is used in for the release channels of fiaas-skipper. There are two
promote stages that can be manually triggered for populating the
different release channels latest/stable with the meta-data of the
current release of fiaas-deploy-daemon.